refactor to implement renderer.NodeRenderer
I've refactored the library to implement goldmark's
renderer.NodeRenderer. Previously it was circumvented with a hacky
solution that shoved all code handling code into a Render function. This
slightly changes how the library is used (the test has been updated),
but it will produce the same output with the default settings.

I wanted to do this both to be more inline with other goldmark renderers
and to allow for more options in the future. You can already see the new
options padded out in options.go -- most are not implemented yet.