implement HeadingLinkOff option
The heading link off option now works. It was a little tricky because
links and autolinks check if their parent is link only and if so won't
print even their label. This means the parent node is responsible for
printing the link labels. I added some comments to clarify this.

Also I fixed #12 while working on the heading function. I twas but a
space typo. Thought it was going to be much more complicated.