create new benchmarks
I've created a benchmark to test the Rendering process rather than the
whole Convert function. Since this library doesn't do any parsing it was
just dirtying the benchmark's output. I also added a Benchmark for
linkOnly in case I want to make some changes to it in the future since
it seems like a possible bottleneck.