config option for horizontal rules
This commit adds a config option for horizontal rules. You can now print
them using any string you like, including newlines to create a codeblock
for example. This commit ALSO changes the default horizontal rule to a
blank string. Having the 80 character - as a default was not ideal for
screenreaders.